Move code to utils module
